+++ design-system/dropdown-menu/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/dropdown-menu
 
+## 11.0.1
+
+### Patch Changes
+
+- [`c2dd770a743`](https://bitbucket.org/atlassian/atlassian-frontend/commits/c2dd770a743) - Add new prop which controls is outside click should be bound using capture
+- Updated dependencies
+
 ## 11.0.0
 
 ### Major Changes